+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed an issue with the Expo plugin where Android extensions would not be properly enabled.
+
 ### Added
 
+- Added `theoads` and `millicast` support for Android Expo plugin.
 - Added support for the new `TextTrackStyle.marginBottom` and `TextTrackStyle.marginRight` API's on THEOplayer Android SDK v9.5.0.
 
 ## [9.3.0] - 25-06-02